Use new format for tests
